+++ src/draw/BoardLayer.ts
@@ -1,4 +1,5 @@
 import { Board } from '../Board';
+import { MODE } from '../Game';
 import { CanvasLayer, DrawingLayer } from './GameDrawer';
 import { Drop } from '../Drop';
 import { PuyoDrawingLayer } from './PuyoDrawingLayer';
@@ -6,18 +7,9 @@ import { Settings } from '../utils/Settings';
 
 import * as CONSTANTS from './DrawingConfig';
 
-enum MODE {
-	PUYO_DROPPING,
-	PUYO_DROPPING_SPLIT,
-	PUYO_SQUISHING,
-	CHAIN_POPPING,
-	CHAIN_DROPPING,
-	CHAIN_SQUISHING,
-	NUISANCE_DROPPING
-}
-
 export class BoardLayer extends CanvasLayer {
-	unit: number;
+	unitW: number;
+	unitH: number;
 	settings: Settings;
 	appearance: string;
 	backgroundLayer: DrawingLayer;
@@ -28,19 +20,24 @@ export class BoardLayer extends CanvasLayer {
 
 	constructor(settings: Settings, appearance: string, scaleFactor = 1, onNode: boolean) {
 		super(CONSTANTS.DIMENSIONS.BOARD.W * scaleFactor, CONSTANTS.DIMENSIONS.BOARD.H * scaleFactor, onNode);
-		this.unit = this.width / settings.cols;
+		this.unitW = this.width / settings.cols;
+		this.unitH = this.height / settings.rows;
 		this.settings = settings;
 		this.appearance = appearance;
-		this.backgroundLayer = new DrawingLayer(this.width, this.height, this.unit, onNode);
+		this.backgroundLayer = new DrawingLayer(this.width, this.height, this.unitW, this.unitH, onNode);
 		this.backgroundLayer.resetState();
 		if (!this.onNode) {
 			this.backgroundLayer.ctx.fillStyle = 'black';
 			this.backgroundLayer.ctx.globalAlpha = 0.2;
 			this.backgroundLayer.ctx.fillRect(0, 0, this.width, this.height);
-			this.backgroundLayer.draw({ appearance, size: 1, sX: CONSTANTS.PUYO_COORDINATES.CROSS.X, sY: CONSTANTS.PUYO_COORDINATES.CROSS.Y, dX: 2.5, dY: 0.5, sWidth: 1, sHeight: 1, merge: false });
+			this.backgroundLayer.draw({
+				appearance, unitW: 1, unitH: 1,
+				sX: CONSTANTS.PUYO_COORDINATES.CROSS.X, sY: CONSTANTS.PUYO_COORDINATES.CROSS.Y,
+				dX: 2.5, dY: 0.5, sWidth: 1, sHeight: 1, merge: false
+			});
 		}
-		this.stackLayer = new PuyoDrawingLayer(this.width, this.height, this.unit, appearance, onNode);
-		this.dynamicLayer = new PuyoDrawingLayer(this.width, this.height, this.unit, appearance, onNode);
+		this.stackLayer = new PuyoDrawingLayer(this.width, this.height, this.unitW, this.unitH, appearance, onNode);
+		this.dynamicLayer = new PuyoDrawingLayer(this.width, this.height, this.unitW, this.unitH, appearance, onNode);
 		this.columnHeights = [];
 		this.mode = null;
 	}
@@ -75,86 +72,106 @@ export class BoardLayer extends CanvasLayer {
 
 	updateBoard(currentBoardState : { connections: Puyo[][], currentDrop: Drop }): void {
 		const { connections, currentDrop } = currentBoardState;
+		const { standardAngle, arle, schezo, colours } = currentDrop;
+		const { rows, cols } = this.settings;
+
 		if (currentDrop.schezo.y == null) {
 			if (this.hasStackChanged(MODE.PUYO_DROPPING)) {
-				this.columnHeights = (new Array(this.settings.cols).fill(0)) as number[];
+				this.columnHeights = new Array<number>(cols).fill(0);
 				this.stackLayer.resetState();
 				connections.forEach(group => {
 					group.forEach(puyo => {
-						this.stackLayer.drawPuyo(puyo.colour, 0.5 + puyo.col, this.settings.rows - 0.5 - puyo.row, puyo.connections);
-						if (this.columnHeights[puyo.col] < puyo.row + 1) {
-							this.columnHeights[puyo.col] = puyo.row + 1;
+						this.stackLayer.drawPuyo(puyo.colour, 0.5 + puyo.x, rows - 0.5 - puyo.y, puyo.connections);
+						if (this.columnHeights[puyo.x] < puyo.y + 1) {
+							this.columnHeights[puyo.x] = puyo.y + 1;
 						}
 					});
 				});
 			}
 			this.dynamicLayer.resetState();
-			if (currentDrop.standardAngle > Math.PI / 4 && currentDrop.standardAngle <= 3 * Math.PI / 4) {
-				this.dynamicLayer.drawGhost(currentDrop.colours[0], 0.5 + currentDrop.arle.x, this.settings.rows - 0.5 - this.columnHeights[currentDrop.arle.x]);
-				if (this.columnHeights[currentDrop.arle.x - 1] <= this.columnHeights[currentDrop.arle.x] || this.columnHeights[currentDrop.arle.x - 1] <= currentDrop.arle.y) {
-					this.dynamicLayer.drawGhost(currentDrop.colours[1], -0.5 + currentDrop.arle.x, this.settings.rows - 0.5 - this.columnHeights[currentDrop.arle.x - 1]);
+			if (standardAngle > Math.PI / 4 && standardAngle <= 3 * Math.PI / 4) {
+				this.dynamicLayer.drawGhost(colours[0], 0.5 + arle.x, rows - 0.5 - this.columnHeights[arle.x]);
+				if (this.columnHeights[arle.x - 1] <= this.columnHeights[arle.x] || this.columnHeights[arle.x - 1] <= arle.y) {
+					this.dynamicLayer.drawGhost(colours[1], -0.5 + arle.x, rows - 0.5 - this.columnHeights[arle.x - 1]);
 				}
 			}
-			else if (currentDrop.standardAngle > 3 * Math.PI / 4 && currentDrop.standardAngle <= 5 * Math.PI / 4) {
-				this.dynamicLayer.drawGhost(currentDrop.colours[1], 0.5 + currentDrop.arle.x, this.settings.rows - 0.5 - this.columnHeights[currentDrop.arle.x]);
-				this.dynamicLayer.drawGhost(currentDrop.colours[0], 0.5 + currentDrop.arle.x, this.settings.rows - 1.5 - this.columnHeights[currentDrop.arle.x]);
+			else if (standardAngle > 3 * Math.PI / 4 && standardAngle <= 5 * Math.PI / 4) {
+				this.dynamicLayer.drawGhost(colours[1], 0.5 + arle.x, rows - 0.5 - this.columnHeights[arle.x]);
+				this.dynamicLayer.drawGhost(colours[0], 0.5 + arle.x, rows - 1.5 - this.columnHeights[arle.x]);
 			}
-			else if (currentDrop.standardAngle > 5 * Math.PI / 4 && currentDrop.standardAngle <= 7 * Math.PI / 4) {
-				this.dynamicLayer.drawGhost(currentDrop.colours[0], 0.5 + currentDrop.arle.x, this.settings.rows - 0.5 - this.columnHeights[currentDrop.arle.x]);
-				if (this.columnHeights[currentDrop.arle.x + 1] <= this.columnHeights[currentDrop.arle.x] || this.columnHeights[currentDrop.arle.x + 1] <= currentDrop.arle.y) {
-					this.dynamicLayer.drawGhost(currentDrop.colours[1], 1.5 + currentDrop.arle.x, this.settings.rows - 0.5 - this.columnHeights[currentDrop.arle.x + 1]);
+			else if (standardAngle > 5 * Math.PI / 4 && standardAngle <= 7 * Math.PI / 4) {
+				this.dynamicLayer.drawGhost(colours[0], 0.5 + arle.x, rows - 0.5 - this.columnHeights[arle.x]);
+				if (this.columnHeights[arle.x + 1] <= this.columnHeights[arle.x] || this.columnHeights[arle.x + 1] <= arle.y) {
+					this.dynamicLayer.drawGhost(colours[1], 1.5 + arle.x, rows - 0.5 - this.columnHeights[arle.x + 1]);
 				}
 			}
 			else {
-				this.dynamicLayer.drawGhost(currentDrop.colours[0], 0.5 + currentDrop.arle.x, this.settings.rows - 0.5 - this.columnHeights[currentDrop.arle.x]);
-				this.dynamicLayer.drawGhost(currentDrop.colours[1], 0.5 + currentDrop.arle.x, this.settings.rows - 1.5 - this.columnHeights[currentDrop.arle.x]);
+				this.dynamicLayer.drawGhost(colours[0], 0.5 + arle.x, rows - 0.5 - this.columnHeights[arle.x]);
+				this.dynamicLayer.drawGhost(colours[1], 0.5 + arle.x, rows - 1.5 - this.columnHeights[arle.x]);
 			}
-			this.dynamicLayer.drawHighlightedDrop(currentDrop, 0.5 + currentDrop.arle.x, this.settings.rows - 0.5 - currentDrop.arle.y);
+			this.dynamicLayer.drawHighlightedDrop(currentDrop, 0.5 + arle.x, rows - 0.5 - arle.y);
 		}
 		else {
 			if (this.hasStackChanged(MODE.PUYO_DROPPING_SPLIT)) {
 				this.stackLayer.resetState();
 				connections.forEach(group => {
 					group.forEach(puyo => {
-						this.stackLayer.drawPuyo(puyo.colour, 0.5 + puyo.col, this.settings.rows - 0.5 - puyo.row, puyo.connections);
+						this.stackLayer.drawPuyo(puyo.colour, 0.5 + puyo.x, rows - 0.5 - puyo.y, puyo.connections);
 					});
 				});
 			}
 			this.dynamicLayer.resetState();
-			this.dynamicLayer.drawPuyo(currentDrop.colours[0], 0.5 + currentDrop.arle.x, this.settings.rows - 0.5 - currentDrop.arle.y);
-			this.dynamicLayer.drawPuyo(currentDrop.colours[1], 0.5 + currentDrop.schezo.x, this.settings.rows - 0.5 - currentDrop.schezo.y);
+			this.dynamicLayer.drawPuyo(colours[0], 0.5 + arle.x, rows - 0.5 - arle.y);
+			this.dynamicLayer.drawPuyo(colours[1], 0.5 + schezo.x, rows - 0.5 - schezo.y);
 		}
 		this.update();
 	}
 
-	squishPuyos(currentBoardState: { connections: Puyo[][], currentDrop: Drop }): void {
-		const { connections } = currentBoardState;
+	squishPuyos(currentBoardState: { connections: Puyo[][], squishingPuyos: { puyo: Puyo, squishType: string }[] }, squishState: { currentFrame: number }): void {
+		const { connections, squishingPuyos } = currentBoardState;
+		// console.log(schezo);
+		const { rows } = this.settings;
 		if (this.hasStackChanged(MODE.PUYO_SQUISHING)) {
 			this.stackLayer.resetState();
 			connections.forEach(group => {
 				group.forEach(puyo => {
-					this.stackLayer.drawPuyo(puyo.colour, 0.5 + puyo.col, this.settings.rows - 0.5 - puyo.row, puyo.connections);
+					this.stackLayer.drawPuyo(puyo.colour, 0.5 + puyo.x, this.settings.rows - 0.5 - puyo.y, puyo.connections);
 				});
 			});
-			this.dynamicLayer.resetState();
+			// this.dynamicLayer.resetState();
+		}
+		this.dynamicLayer.resetState();
+		if(squishingPuyos == null) {
+			console.log(squishingPuyos);
+			return;
+		}
+		for(const puyo of squishingPuyos) {
+			const { colour, x, y } = puyo.puyo;
+			const type = CONSTANTS.SQUISH_FRAMES[puyo.squishType][Math.floor(squishState.currentFrame / 2)];
+			if(type === undefined) {
+				continue;
+			}
+			this.dynamicLayer.drawSquishingPuyo(colour, 0.5 + x, rows - 0.5 - y, type);
 		}
 		this.update();
 	}
 
 	resolveChains(resolvingState: ResolvingState): void {
-		const { connections, poppedLocs, connectionsAfterPop, unstablePuyos } = resolvingState;
-		if (resolvingState.currentFrame <= this.settings.popFrames) {
+		const { connections, poppedLocs, connectionsAfterPop, unstablePuyos, currentFrame } = resolvingState;
+		const { rows, popFrames, dropFrames } = this.settings;
+
+		if (currentFrame <= popFrames) {
 			if (this.hasStackChanged(MODE.CHAIN_POPPING)) {
 				this.stackLayer.resetState();
 				connections.forEach(group => {
-					group.filter(puyo => !poppedLocs.some(puyo2 => puyo.col === puyo2.col && puyo.row === puyo2.row)).forEach(puyo => {
-						this.stackLayer.drawPuyo(puyo.colour, 0.5 + puyo.col, this.settings.rows - 0.5 - puyo.row, puyo.connections);
+					group.filter(puyo => !poppedLocs.some(puyo2 => puyo.x === puyo2.x && puyo.y === puyo2.y)).forEach(puyo => {
+						this.stackLayer.drawPuyo(puyo.colour, 0.5 + puyo.x, rows - 0.5 - puyo.y, puyo.connections);
 					});
 				});
 			}
 			this.dynamicLayer.resetState();
-			poppedLocs.forEach(puyo => {
-				this.dynamicLayer.drawPoppingPuyo(puyo.colour, 0.5 + puyo.col, this.settings.rows - 0.5 - puyo.row, resolvingState.currentFrame >= this.settings.popFrames / 3);
+			poppedLocs.forEach((puyo, index) => {
+				this.dynamicLayer.drawPoppingPuyo(puyo.colour, 0.5 + puyo.x, rows - 0.5 - puyo.y, puyo.connections, currentFrame, index);
 			});
 		}
 		else {
@@ -162,16 +179,16 @@ export class BoardLayer extends CanvasLayer {
 				this.stackLayer.resetState();
 				connectionsAfterPop.forEach(group => {
 					group.forEach(puyo => {
-						this.stackLayer.drawPuyo(puyo.colour, 0.5 + puyo.col, this.settings.rows - 0.5 - puyo.row, puyo.connections);
+						this.stackLayer.drawPuyo(puyo.colour, 0.5 + puyo.x, rows - 0.5 - puyo.y, puyo.connections);
 					});
 				});
 			}
 			this.dynamicLayer.resetState();
-			unstablePuyos.filter(puyo => !poppedLocs.some(puyo2 => puyo.col === puyo2.col && puyo.row === puyo2.row)).forEach(puyo => {
+			unstablePuyos.filter(puyo => !poppedLocs.some(puyo2 => puyo.x === puyo2.x && puyo.y === puyo2.y)).forEach(puyo => {
 				this.dynamicLayer.drawPuyo(
 					puyo.colour,
-					0.5 + puyo.col,
-					this.settings.rows - 0.5 - Math.max(puyo.row - (puyo.row - puyo.above) * (resolvingState.currentFrame - this.settings.popFrames) / this.settings.dropFrames, puyo.above)
+					0.5 + puyo.x,
+					this.settings.rows - 0.5 - Math.max(puyo.y - (puyo.y - puyo.above) * (currentFrame - popFrames) / dropFrames, puyo.above)
 				);
 			});
 		}
@@ -185,7 +202,7 @@ export class BoardLayer extends CanvasLayer {
 			const connections = new Board(this.settings, boardState).getConnections();
 			connections.forEach(group => {
 				group.forEach(puyo => {
-					this.stackLayer.drawPuyo(puyo.colour, 0.5 + puyo.col, this.settings.rows - 0.5 - puyo.row, puyo.connections);
+					this.stackLayer.drawPuyo(puyo.colour, 0.5 + puyo.x, this.settings.rows - 0.5 - puyo.y, puyo.connections);
 				});
 			});
 		}