+++ src/draw/PuyoDrawingLayer.ts
@@ -3,11 +3,12 @@ import { Drop } from '../Drop';
 import * as CONSTANTS from './DrawingConfig';
 
 export class PuyoDrawingLayer extends DrawingLayer {
-	constructor(width: number, height: number, unit: number, appearance: string, onNode: boolean, className = undefined) {
-		super(width, height, unit, onNode, className);
+	constructor(width: number, height: number, unitW: number, unitH: number, appearance: string, onNode: boolean, className = undefined) {
+		super(width, height, unitW, unitH, onNode, className);
 		this.defaultArgs = {
 			appearance,
-			size: 1,
+			unitW: this.unitW,
+			unitH: this.unitH,
 			sWidth: 1,
 			sHeight: 1,
 			merge: true
@@ -39,8 +40,42 @@ export class PuyoDrawingLayer extends DrawingLayer {
 		this.draw({ sX, sY, dX, dY });
 	}
 
-	drawPoppingPuyo(colour: number, dX: number, dY: number, drawPhaseTwo: boolean): void {
-		const { X: sX, Y: sY } = CONSTANTS.PUYO_COORDINATES.POPPING[colour][drawPhaseTwo ? 2 : 1];
+	drawPoppingPuyo(colour: number, dX: number, dY: number, directions: string[] = [], currentFrame: number, offset: number): void {
+		if(colour === 0) {
+			return;
+		}
+		currentFrame = Math.floor(currentFrame / 2);
+		if(currentFrame <= CONSTANTS.POPPING_FRAMES) {
+			// FLicker on and off
+			if(currentFrame % 2 === 0) {
+				this.drawPuyo(colour, dX, dY, directions);
+			}
+		}
+		else {
+			let sX: number, sY: number, scale: number;
+			const index = currentFrame - CONSTANTS.POPPING_FRAMES - offset;
+
+			if(index < 5) {
+				({ X: sX, Y: sY } = CONSTANTS.PUYO_COORDINATES.POPPING[colour][1]);
+				this.draw({ sX, sY, dX, dY });
+			}
+			else if (index < 7){
+				({ X: sX, Y: sY } = CONSTANTS.PUYO_COORDINATES.POPPING[colour][index - 3]);
+				this.draw({ sX, sY, dX, dY });
+			}
+
+			if(index > 4) {
+				({ X: sX, Y: sY, SCALE: scale } = CONSTANTS.PUYO_COORDINATES.EXPLOSION[colour][0]);
+				const x = (index - 6) / 4;
+				const y = Math.pow(x - 1, 2) - 1;
+				this.draw({ sX, sY, dX: dX + x, dY: dY + y, sWidth: scale, sHeight: scale});
+				this.draw({ sX, sY, dX: dX - x, dY: dY + y, sWidth: scale, sHeight: scale});
+			}
+		}
+	}
+
+	drawSquishingPuyo(colour: number, dX: number, dY: number, type: SquishType): void {
+		const { X: sX, Y: sY } = CONSTANTS.PUYO_COORDINATES.SQUISHING[colour][type];
 		this.draw({ sX, sY, dX, dY });
 	}
 
@@ -81,11 +116,17 @@ export class PuyoDrawingLayer extends DrawingLayer {
 }
 
 export class QueueLayer extends PuyoDrawingLayer {
-	updateQueue(queueState: { dropArray: Drop[] }): void {
-		const { dropArray } = queueState;
+	lastDrop: Drop = null;
+
+	updateQueue(queueState: { dropArray: Drop[], currentFrame: number }): void {
+		const { dropArray, currentFrame } = queueState;
 		this.resetState();
-		for (let i = 0; i < dropArray.length; i++) {
-			this.drawDrop(dropArray[i], 1, 2 + 3 * i);
+		if(this.lastDrop !== null) {
+			this.drawDrop(this.lastDrop, 1, 2 - 3 * (currentFrame / 4));
 		}
+		for (let i = 0; i < dropArray.length; i++) {
+			this.drawDrop(dropArray[i], 1, 5 + 3 * (i - currentFrame / 4));
+		}
+		this.lastDrop = dropArray[0];
 	}
 }